# docs/api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GOF API",
        "description": "API REST do GOF (Gestão Operacional de Frotas). Todas as respostas, inclusive erros, usam o envelope {status, message, errors, data[, meta]}.",
        "license": {
            "name": "Proprietary"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1 (mesma origem, via nginx)"
        }
    ],
    "paths": {
...
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login (emite token Bearer)",
                "description": "Público. `login` com `@` é e-mail; sem `@` é username (os dois com trim + minúsculas). Credenciais inválidas → 422 em `errors.login` com mensagem genérica, igual para e-mail e username (não revela se o login existe). Usuário inativo → 403 (somente com a senha correta). Rate limit: 5/min por login normalizado + IP e 20/min por IP (toda tentativa conta). Token expira em SANCTUM_EXPIRATION minutos (padrão 720).",
                "operationId": "authLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password",
                                    "device_name"
                                ],
                                "properties": {
                                    "login": {
                                        "description": "E-mail ou username.",
                                        "type": "string",
                                        "example": "admin@example.com",
                                        "maxLength": 190
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "maxLength": 255
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "web",
                                        "maxLength": 255
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Autenticado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Envelope"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/LoginData"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": []
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout (revoga o token atual)",
                "operationId": "authLogout",
                "responses": {
                    "200": {
                        "description": "Token revogado (`data` null).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Envelope"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Usuário autenticado, colaborador vinculado e permissões",
                "operationId": "authMe",
                "responses": {
                    "200": {
                        "description": "Dados da sessão.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Envelope"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/MeData"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/password": {
            "put": {
                "tags": [
                    "Auth"
                ],
                "summary": "Troca a própria senha (revoga os demais tokens)",
                "description": "Política: mín. 10 caracteres, com maiúscula, minúscula e número. O token usado na requisição continua válido; todos os outros do usuário são revogados.",
                "operationId": "authChangePassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "maxLength": 255,
                                        "minLength": 10
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Senha alterada (`data` null).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Envelope"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthenticated"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/forgot-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Pede o link de redefinição de senha",
                "description": "Público. Sempre 200 com a mesma mensagem, exista ou não o e-mail (sem enumeração). Só usuário ativo e não excluído recebe o e-mail, enviado pela fila; o link é `${APP_FRONTEND_URL}/redefinir-senha?token=…&email=…`, vale 60 min e um pedido novo invalida o anterior. E-mail vazio ou malformado → 422. Rate limit: 3/min por e-mail + IP e 10/min por IP.",
                "operationId": "authForgotPassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ana@example.com",
                                        "maxLength": 190
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mensagem genérica (`data` null).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Envelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": []
            }
        },
        "/auth/reset-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Redefine a senha com o link recebido por e-mail",
                "description": "Público. Política de senha → 422 por campo (`password`, inclusive a confirmação). Link inválido, expirado, já usado, de outro e-mail ou de usuário inativo/excluído → 422 genérico em `errors.token`. Sucesso revoga todos os tokens de acesso do usuário e não faz login automático. Rate limit: 5/min por IP.",
                "operationId": "authResetPassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "token",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 190
                                    },
                                    "token": {
                                        "description": "Token do link do e-mail.",
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "maxLength": 255,
                                        "minLength": 10
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Senha redefinida (`data` null).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Envelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "429": {
                        "$ref": "#/components/responses/TooManyRequests"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": []
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LoginData": {
                "required": [
                    "token",
                    "token_type",
                    "expires_at",
                    "user"
                ],
                "properties": {
                    "token": {
                        "type": "string",
                        "example": "1|9fQk..."
                    },
                    "token_type": {
                        "type": "string",
                        "enum": [
                            "Bearer"
                        ]
                    },
                    "expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-10-09T02:03:22.123456Z"
                    },
                    "user": {
                        "$ref": "#/components/schemas/AuthUser"
                    }
                },
                "type": "object"
            },
...
            "Envelope": {
                "description": "Envelope padrão de toda resposta.",
                "required": [
                    "status",
                    "message",
                    "errors",
                    "data"
                ],
                "properties": {
                    "status": {
                        "description": "Definições globais da OpenAPI (info, servidor, segurança, schemas e respostas reutilizáveis).\nOs endpoints são anotados nos próprios controllers. O atributo não aceita expressão, então\ntítulo e descrição usam os marcadores de OpenApiBuilder::BRAND_TOKENS, preenchidos com a\nmarca do config (app.name / app.full_name) na geração.",
                        "type": "string",
                        "enum": [
                            "success",
                            "error"
                        ]
                    },
                    "message": {
                        "type": "string",
                        "example": "Operação realizada com sucesso."
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "data": {
                        "description": "Objeto, lista ou null conforme o endpoint (qualquer tipo)."
                    },
                    "meta": {
                        "$ref": "#/components/schemas/Pagination"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token obtido em POST /auth/login (expira em 12 h).",
                "bearerFormat": "Sanctum personal access token",
                "scheme": "bearer"
            }
        }
    },
...
}